NewSignedUpdater to updater
