+++ repo/pullrequest_test.go
@@ -2,9 +2,6 @@ package repo_test
 
 import (
 	"context"
-	"encoding/base64"
-	"encoding/json"
-	"fmt"
 	"strings"
 	"testing"
 
@@ -98,47 +95,4 @@ Here are some updates, I hope they work.
 {"updates":[{"path":"github.com/aws/aws-sdk-go","previous":"v1.34.16","next":"v1.34.17"},{"path":"github.com/foo/bar","previous":"v0.4.1","next":"v99.88.77"}],"signature":"TL6d3v5DKRu8uDY5doooDLLd7mJDHx6U5P4jRZYanLT4VI1dzt1gIRvZGW3G0ZlDQqmuTOftovTlwLHO1VW4Xw=="}
 -->
 `), strings.TrimSpace(body))
-}
-
-func TestNewSignedUpdateDescriptor(t *testing.T) {
-	cases := []struct {
-		signature string
-		updates   []updater.Update
-	}{
-		{
-			signature: "HAF6zSdBBOsbrLRClce7M73tN7VhCdPB6YYhECL/ifDC6DHR0YSGXoY6JQeEaFoncJbxp/afBpY+GVE5DUfWwQ==",
-			updates:   []updater.Update{awsSdkGo13417},
-		},
-		{
-			signature: "kq9CbO3rYkThJPiJgVTfhkfAG4q5aEeXuta0x3wPVdUnqQhitA/FasfJ2WftpfiZvueCnknoX04yxTM94BUn4A==",
-			updates:   []updater.Update{fooBar987},
-		},
-		{
-			signature: "TL6d3v5DKRu8uDY5doooDLLd7mJDHx6U5P4jRZYanLT4VI1dzt1gIRvZGW3G0ZlDQqmuTOftovTlwLHO1VW4Xw==",
-			updates:   []updater.Update{awsSdkGo13417, fooBar987},
-		},
-	}
-
-	for _, tc := range cases {
-		t.Run(fmt.Sprintf("%v", tc.updates), func(t *testing.T) {
-			descriptor, err := repo.NewSignedUpdateDescriptor(testKey, tc.updates...)
-			require.NoError(t, err)
-
-			buf, err := json.Marshal(&descriptor)
-			require.NoError(t, err)
-			t.Log(string(buf))
-
-			assert.Equal(t, tc.updates, descriptor.Updates)
-			assert.Equal(t, tc.signature, base64.StdEncoding.EncodeToString(descriptor.Signature))
-
-			verified, err := repo.VerifySignedUpdateDescriptor(testKey, descriptor)
-			require.NoError(t, err)
-			assert.Equal(t, tc.updates, verified)
-		})
-	}
-}
-
-func TestVerifySignedUpdateDescriptor_Invalid(t *testing.T) {
-	_, err := repo.VerifySignedUpdateDescriptor([]byte{}, repo.SignedUpdateDescriptor{})
-	assert.EqualError(t, err, "invalid signature")
 }
